INT-1246 move CodemodListProvider into the MainProvider
